100detoursBack/requirements.txt
flask
flask_accepts==0.15.5
flask-cors
flask-mongoengine
flask-restplus
flask-jwt-extended
marshmallow
passlib
requests